requirements.txt
# ============================================================================
# LIBRA ENERGIA - DEPENDÊNCIAS
# ============================================================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlite3  # Built-in with Python

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
webdriver-manager==4.0.1
fake-useragent==1.4.0

# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Google Sheets
gspread==5.12.0
oauth2client==4.1.3

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Environment Variables
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2

# JSON Processing
jsonschema==4.20.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging
loguru==0.7.2

# CORS
python-multipart==0.0.6

# Validation
pydantic==2.5.0
email-validator==2.1.0

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Monitoring
prometheus-client==0.19.0

# Optional: For production deployment
# gunicorn==21.2.0
# psycopg2-binary==2.9.9  # PostgreSQL
# redis==5.0.1  # Caching
# celery==5.3.4  # Task queue